+++ Module_4/T1-mammals/src/dog.cpp
@@ -1,5 +1,6 @@
 #include "dog.hpp"
 
-Dog::Dog(const std::string &name, int weight) : Mammal(weight), name_(name) {}
+Dog::Dog(const std::string &name, double weight)
+    : Mammal(weight), name_(name) {}
 const std::string &Dog::GetName() const { return name_; }
 std::string Dog::MakeSound() { return "Wuff!"; }
